refactor: move ecstore pool-space helpers
