+++ crates/ecstore/src/store.rs
@@ -87,7 +87,6 @@ use s3s::dto::{BucketVersioningStatus, ObjectLockConfiguration, ObjectLockEnable
 use std::cmp::Ordering;
 use std::net::SocketAddr;
 use std::process::exit;
-use std::slice::Iter;
 use std::time::SystemTime;
 use std::{
     collections::HashMap,
@@ -106,6 +105,8 @@ type ListObjectsV2Info = StorageListObjectsV2Info<ObjectInfo>;
 type ListObjectVersionsInfo = StorageListObjectVersionsInfo<ObjectInfo>;
 type ObjectInfoOrErr = StorageObjectInfoOrErr<ObjectInfo, Error>;
 type WalkOptions = StorageWalkOptions<fn(&FileInfo) -> bool>;
+
+pub use crate::layout::pool_space::{PoolAvailableSpace, ServerPoolsAvailableSpace};
 
 /// Check if a directory contains any xl.meta files (indicating actual S3 objects)
 /// This is used to determine if a bucket is empty for deletion purposes.
@@ -791,59 +792,6 @@ impl rustfs_storage_api::StorageAdminApi for ECStore {
     }
 }
 
-#[derive(Debug, Default, Clone)]
-pub struct PoolAvailableSpace {
-    pub index: usize,
-    pub available: u64,    // in bytes
-    pub max_used_pct: u64, // Used disk percentage of most filled disk, rounded down.
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct ServerPoolsAvailableSpace(Vec<PoolAvailableSpace>);
-
-impl ServerPoolsAvailableSpace {
-    pub fn iter(&self) -> Iter<'_, PoolAvailableSpace> {
-        self.0.iter()
-    }
-    // TotalAvailable - total available space
-    pub fn total_available(&self) -> u64 {
-        let mut total = 0;
-        for pool in &self.0 {
-            total += pool.available;
-        }
-        total
-    }
-
-    // FilterMaxUsed will filter out any pools that has used percent bigger than max,
-    // unless all have that, in which case all are preserved.
-    pub fn filter_max_used(&mut self, max: u64) {
-        if self.0.len() <= 1 {
-            // Nothing to do.
-            return;
-        }
-        let mut ok = false;
-        for pool in &self.0 {
-            if pool.available > 0 && pool.max_used_pct < max {
-                ok = true;
-                break;
-            }
-        }
-        if !ok {
-            // All above limit.
-            // Do not modify
-            return;
-        }
-
-        // Remove entries that are above.
-        for pool in self.0.iter_mut() {
-            if pool.available > 0 && pool.max_used_pct < max {
-                continue;
-            }
-            pool.available = 0
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -870,32 +818,6 @@ mod tests {
         let result = has_space_for(&disk_infos, 1024).await;
         // Should fail due to no valid disk info
         assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_server_pools_available_space() {
-        let mut spaces = ServerPoolsAvailableSpace(vec![
-            PoolAvailableSpace {
-                index: 0,
-                available: 1000,
-                max_used_pct: 50,
-            },
-            PoolAvailableSpace {
-                index: 1,
-                available: 2000,
-                max_used_pct: 80,
-            },
-        ]);
-
-        assert_eq!(spaces.total_available(), 3000);
-
-        spaces.filter_max_used(60);
-        // filter_max_used sets available to 0 for filtered pools, doesn't remove them
-        assert_eq!(spaces.0.len(), 2); // Length remains the same
-        assert_eq!(spaces.0[0].index, 0);
-        assert_eq!(spaces.0[0].available, 1000); // First pool should still be available
-        assert_eq!(spaces.0[1].available, 0); // Second pool should be filtered (available = 0)
-        assert_eq!(spaces.total_available(), 1000); // Only first pool contributes to total
     }
 
     #[tokio::test]
@@ -994,34 +916,5 @@ mod tests {
         };
 
         assert!(!should_enqueue_transition_immediately(&oi));
-    }
-
-    // Test that we can create the basic structures without global state
-    #[test]
-    fn test_pool_available_space_creation() {
-        let space = PoolAvailableSpace {
-            index: 0,
-            available: 1000,
-            max_used_pct: 50,
-        };
-        assert_eq!(space.index, 0);
-        assert_eq!(space.available, 1000);
-        assert_eq!(space.max_used_pct, 50);
-    }
-
-    #[test]
-    fn test_server_pools_available_space_iter() {
-        let spaces = ServerPoolsAvailableSpace(vec![PoolAvailableSpace {
-            index: 0,
-            available: 1000,
-            max_used_pct: 50,
-        }]);
-
-        let mut count = 0;
-        for space in spaces.iter() {
-            assert_eq!(space.index, 0);
-            count += 1;
-        }
-        assert_eq!(count, 1);
     }
 }
